css/selectors/focus-visible-007-manual.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <title>CSS Test (Selectors): Keyboard focus enables :focus-visible</title>
  <link rel="author" title="Alice Boxhall" href="aboxhall@chromium.org" />
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <style>
    body[data-hadkeydown] #one:focus-visible { background-color: rgb(128, 196, 128); }
    body:not([data-hadkeydown]) #one:focus-visible { background-color: rgb(196, 128, 128); }

    body:not([data-hadkeydown]) #two:focus-visible { background-color: rgb(196, 128, 128); }
    body:not([data-hadkeydown]) #two:focus:not(:focus-visible) {
      background-color: rgb(128, 196, 128);
    }
  </style>
</head>
<body>
  <ol id="instructions">
    <li>If the user-agent does not claim to support the <code>:focus-visible</code> pseudo-class then SKIP this test.</li>
    <li>Use the mouse to focus the element below that says "Click me first."</li>
    <li>If the element has a red background, then the test result is FAILURE.</li>
    <li>Press the SHIFT key.</li>
    <li>If the element now DOES NOT have a green background, then the test result is FAILURE..</li>
    <li>Use the mouse to click the element that says "Click me first." a second time</li>
    <li>If the element now DOES NOT still have a green background, then the test result is FAILURE.</li>
    <li>Use the mouse to click the element below that says "Click me second."</li>
    <li>If the element has a green background then the test result is SUCCESS.</li>
  </ol>
  <br />
  <div id="one" tabindex="0">Click me first.</div>
  <div id="two" tabindex="0">Click me second.</div>
  <script>
    document.body.addEventListener("keydown", (e) => {
      delete document.body.dataset.hadmousedown;
      document.body.dataset.hadkeydown = "";
    }, true);

    document.body.addEventListener("mousedown", (e) => {
      document.body.dataset.hadmousedown = "";
    }, true);

    document.body.addEventListener("focus", (e) => {
      if ("hadmousedown" in document.body.dataset) {
        delete document.body.dataset.hadmousedown;
        delete document.body.dataset.hadkeydown;
      }
    }, true);

    async_test(function(t) {
      let tested_modality_change = false;
      let tested_modality_unchanged_by_mouse_click = false;
      let tested_mouse_focus_after_modality_change = false;

      const handle_initial_focus = t.step_func(() => {
        one.addEventListener("keyup", t.step_func(test_modality_change));
        one.removeEventListener("focus", handle_initial_focus);

        // ** REVIEWERS PLEASE READ **
        // This is illegal in WPT tests, but there doesn't seem to be another way to test
        // both keyboard *and* mouse usage in one test, which is critical here.
        // Suggestions welcome!
        eventSender.keyDown("ShiftLeft");
      });

      const test_modality_change = t.step_func(() => {
        assert_equals(getComputedStyle(one).backgroundColor, "rgb(128, 196, 128)");
        tested_modality_change = true;
        one.removeEventListener("keyup", test_modality_change);
        one.addEventListener("click", test_modality_unchanged_by_mouse_click);
        mouseClickInTarget("#one");
      });

      const test_modality_unchanged_by_mouse_click = t.step_func(() => {
        assert_true(tested_modality_change);
        assert_equals(getComputedStyle(one).backgroundColor, "rgb(128, 196, 128)");
        tested_modality_unchanged_by_mouse_click = true;
        one.removeEventListener("click", test_modality_unchanged_by_mouse_click);
        two.addEventListener("focus", test_mouse_focus_after_modality_change);
        mouseClickInTarget("#two");
      });

      const test_mouse_focus_after_modality_change = t.step_func(() => {
        assert_true(tested_modality_unchanged_by_mouse_click);
        assert_false("hadkeydown" in document.body.dataset);
        assert_equals(getComputedStyle(two).backgroundColor, "rgb(128, 196, 128)");
        tested_mouse_focus_after_modality_change = true;
        t.done();
      });

      one.addEventListener("focus", handle_initial_focus);
    }, "Using keyboard while element is focused should trigger :focus-visible; using mouse without moving focus should not cancel it; moving focus using mouse should cancel it.");
  </script>
</body>
</html>
